+++ stanza/id.go
@@ -66,7 +66,8 @@ func (id OriginID) WriteXML(w xmlstream.TokenWriter) (int, error) {
 	return xmlstream.Copy(w, id.TokenReader())
 }
 
-func isStanza(name xml.Name) bool {
+// Is tests whether name is a valid stanza based on the localname and namespace.
+func Is(name xml.Name) bool {
 	return (name.Local == "iq" || name.Local == "message" || name.Local == "presence") &&
 		(name.Space == ns.Client || name.Space == ns.Server)
 }
@@ -75,7 +76,7 @@ func isStanza(name xml.Name) bool {
 // does not already have one.
 func AddID(by jid.JID) xmlstream.Transformer {
 	return xmlstream.InsertFunc(func(start xml.StartElement, level uint64, w xmlstream.TokenWriter) error {
-		if isStanza(start.Name) && level == 1 {
+		if Is(start.Name) && level == 1 {
 			_, err := ID{
 				ID: attr.RandomLen(idLen),
 				By: by,
@@ -88,7 +89,7 @@ func AddID(by jid.JID) xmlstream.Transformer {
 
 var (
 	addOriginID = xmlstream.InsertFunc(func(start xml.StartElement, level uint64, w xmlstream.TokenWriter) error {
-		if isStanza(start.Name) && level == 1 {
+		if Is(start.Name) && level == 1 {
 			_, err := OriginID{
 				ID: attr.RandomLen(idLen),
 			}.WriteXML(w)