+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to this project will be documented in this file.
 - stanza: new functions `AddID` and `AddOriginID` to support unique and stable
   stanza IDs
 - stanza: ability to compare errors with `errors.Is`
+- stanza: add `Is` function to check if an XMLName is a valid stanza name
 - styling: satisfy `fmt.Stringer` for the `Style` type
 - version: new package implementing [XEP-0092: Software Version]
 - xmpp: satisfy `fmt.Stringer` for the `SessionState` type